+++ lua/lazyvim/plugins/extras/util/dot.lua
@@ -10,26 +10,27 @@ return {
 
   -- Add Hyprland Parser
   {
-    "luckasRanarison/tree-sitter-hypr",
+    "luckasRanarison/tree-sitter-hyprlang",
+    dependencies = {
+      "nvim-treesitter/nvim-treesitter",
+      opts = function(_, opts)
+        require("nvim-treesitter.parsers").get_parser_configs().hyprlang = {
+          install_info = {
+            url = "https://github.com/luckasRanarison/tree-sitter-hyprlang",
+            files = { "src/parser.c" },
+            branch = "master",
+          },
+          filetype = "hyprlang",
+        }
+
+        opts.ensure_installed = opts.ensure_installed or {}
+        vim.list_extend(opts.ensure_installed, { "hyprlang" })
+      end,
+    },
     enabled = function()
       return have("hypr")
     end,
     event = "BufRead */hypr/*.conf",
-    build = ":TSUpdate hypr",
-    config = function()
-      -- Fix ft detection for hyprland
-      vim.filetype.add({
-        pattern = { [".*/hypr/.*%.conf"] = "hypr" },
-      })
-      require("nvim-treesitter.parsers").get_parser_configs().hypr = {
-        install_info = {
-          url = "https://github.com/luckasRanarison/tree-sitter-hypr",
-          files = { "src/parser.c" },
-          branch = "master",
-        },
-        filetype = "hypr",
-      }
-    end,
   },
 
   -- add some stuff to treesitter