+++ lua/lazyvim/plugins/lsp/init.lua
@@ -26,6 +26,14 @@ return {
           -- prefix = "icons",
         },
         severity_sort = true,
+        signs = {
+          text = {
+            [vim.diagnostic.severity.ERROR] = require("lazyvim.config").icons.diagnostics.Error,
+            [vim.diagnostic.severity.WARN] = require("lazyvim.config").icons.diagnostics.Warn,
+            [vim.diagnostic.severity.HINT] = require("lazyvim.config").icons.diagnostics.Hint,
+            [vim.diagnostic.severity.INFO] = require("lazyvim.config").icons.diagnostics.Info,
+          },
+        },
       },
       -- Enable this to enable the builtin LSP inlay hints on Neovim >= 0.10.0
       -- Be aware that you also will need to properly configure your LSP server to
@@ -86,7 +94,7 @@ return {
       -- setup autoformat
       Util.format.register(Util.lsp.formatter())
 
-      -- deprectaed options
+      -- deprecated options
       if opts.autoformat ~= nil then
         vim.g.autoformat = opts.autoformat
         Util.deprecate("nvim-lspconfig.opts.autoformat", "vim.g.autoformat")